swift backend: fix premature release of semaphore in Load() & document concurrency issue in List().
refactor wrapReader from b2 -> semaphore so it can be used elsewhere.

As per discussion in PR #1399.
